requirements.txt
Django==5.2.13
djangorestframework==3.17.1
drf-spectacular==0.29.0
flake8==7.3.0
pillow==12.2.0
python-dotenv==1.2.2
djangorestframework-simplejwt=5.5.1
